battle1: the camera publisher and the angle nudge; park six near-misses
func_800BC754 and func_800B45F0 MATCH. `make build` 19x OK,
`tools/audit_parked.sh src/battle/battle1.c` clean on all six parked bodies.

The finding is the one that makes the rest of this unit cheap, and it is
recorded in CLAUDE.md against the existing backward-goto bullet: **read the
absence of hoisting, not the duplicated exit test.** A counted walk whose
target rebuilds a symbol address every iteration, re-reads an obviously
invariant bound, keeps a sign-extension inside the body and computes
`i * stride` per iteration where you emit a walking pointer is a backward
`goto` -- and all four tells appear together, so one reading of the diff
settles it. Three functions here measured 39 -> 14, 32 -> 27 and 29 -> 27 rows
on the loop spelling alone, and func_800B45F0 went on to match once its
D_800E8F94 base was pulled into a named local assigned ahead of the
arithmetic rather than inside the pointer expression.

Parked, each with its row count, its length and its rejected spellings:

  func_800BA2BC  13 rows, exact length -- allocno naming, and cc1 -dl shows
                 the winning quantity already has the higher priority, so
                 neither term is reachable from C. Not a permuter target.
  func_800B5C1C   9 rows, exact length -- 6 of them splat naming plain
                 constants as func_801B0000 / D_801AFFC0 / D_801AFFC4, which
                 are byte-identical; 3 real, a $v0/$v1 tie in block_alloc.
  func_800BB430  18 rows, -1
  func_800B5FE8  20 rows, -2
  func_800B8A34  18 rows, +4 (four load-delay nops; both sides of the
                 MEM_IN_STRUCT_P lever measured and closed)
  func_800C5694  15 rows, +1

`tools/audit_parked.sh` now scans src/battle/battle1.c by default.

+++ src/battle/battle_private.h
@@ -598,7 +598,7 @@ void func_800B6B98(s32, s32);
 void func_800B7FB4(void);
 s16 func_800B888C(s32);
 void func_800B8438(void);
-void func_800B8A34(s16, s32);
+void func_800B8A34(s32, s32);
 static void func_800BA40C();
 static void func_800BB030(s16 arg0);
 void func_800BB2A8(u8);
